- 调整打包配置 - 命名导出IoC和BasePlugin
