+++ demo/index.html
@@ -10,7 +10,8 @@
             packages: [
                 {
                     name: 'ioc',
-                    location: '../src'
+                    location: '../dist',
+                    main: 'bundle'
                 }
             ]
         });
