+++ package.json
@@ -1,65 +1,64 @@
 {
-  "name": "uioc",
-  "version": "1.0.0-alpha.1",
-  "description": "an ioc framework",
-  "main": "dist/bundle.js",
-  "maintainers": {
-    "name": "Exodia",
-    "email": "d_xinxin@163.com"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/ecomfe/uioc.git"
-  },
-  "keywords": [
-    "ioc",
-    "di"
-  ],
-  "scripts": {
-    "test": "karma start --single-run",
-    "start": "karma start",
-    "build": "node scripts/build.js",
-    "doc": "esdoc -c esdoc.json"
-  },
-  "author": {
-    "name": "Exodia",
-    "email": "d_xinxin@163.com"
-  },
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/ecomfe/uioc/issues"
-  },
-  "homepage": "https://github.com/ecomfe/uioc",
-  "devDependencies": {
-    "babel-plugin-transform-es2015-modules-umd": "^6.6.5",
-    "babel-polyfill": "^6.7.2",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-es2015-rollup": "^1.1.1",
-    "babel-preset-stage-0": "^6.5.0",
-    "esdoc": "^0.4.7",
-    "esdoc-es7-plugin": "0.0.3",
-    "karma": "^0.12.16",
-    "karma-babel-preprocessor": "^6.0.1",
-    "karma-chrome-launcher": "^0.1.4",
-    "karma-coverage": "^0.2.4",
-    "karma-firefox-launcher": "^0.1.7",
-    "karma-ie-launcher": "^0.1.5",
-    "karma-jasmine": "^0.2.3",
-    "karma-phantomjs-launcher": "^0.1.4",
-    "karma-requirejs": "^0.2.6",
-    "karma-safari-launcher": "^0.1.1",
-    "karma-sourcemap-loader": "^0.3.7",
-    "phantomjs": "^2.1.7",
-    "requirejs": "^2.2.0",
-    "rollup": "^0.25.8",
-    "rollup-plugin-babel": "^2.4.0",
-    "rollup-plugin-uglify": "^0.3.1",
-    "systemjs-builder": "^0.15.14"
-  },
-  "dependencies": {
-    "uaop": "0.0.1-alpha.2"
-  },
-  "edp": {
-    "main": "main"
-  }
+    "name": "uioc",
+    "version": "1.0.0-alpha.1",
+    "description": "an ioc framework",
+    "main": "dist/bundle.js",
+    "maintainers": {
+        "name": "Exodia",
+        "email": "d_xinxin@163.com"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git://github.com/ecomfe/uioc.git"
+    },
+    "keywords": [
+        "ioc",
+        "di"
+    ],
+    "scripts": {
+        "test": "karma start --single-run",
+        "start": "karma start",
+        "build": "node scripts/build.js",
+        "doc": "esdoc -c esdoc.json"
+    },
+    "author": {
+        "name": "Exodia",
+        "email": "d_xinxin@163.com"
+    },
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/ecomfe/uioc/issues"
+    },
+    "homepage": "https://github.com/ecomfe/uioc",
+    "devDependencies": {
+        "babel-plugin-transform-es2015-modules-umd": "^6.6.5",
+        "babel-polyfill": "^6.7.2",
+        "babel-preset-es2015": "^6.6.0",
+        "babel-preset-es2015-rollup": "^1.1.1",
+        "babel-preset-stage-0": "^6.5.0",
+        "esdoc": "^0.4.7",
+        "esdoc-es7-plugin": "0.0.3",
+        "karma": "^0.12.16",
+        "karma-babel-preprocessor": "^6.0.1",
+        "karma-chrome-launcher": "^0.1.4",
+        "karma-coverage": "^0.2.4",
+        "karma-firefox-launcher": "^0.1.7",
+        "karma-ie-launcher": "^0.1.5",
+        "karma-jasmine": "^0.2.3",
+        "karma-phantomjs-launcher": "^0.1.4",
+        "karma-requirejs": "^0.2.6",
+        "karma-safari-launcher": "^0.1.1",
+        "karma-sourcemap-loader": "^0.3.7",
+        "phantomjs-prebuilt": "^2.1.7",
+        "requirejs": "^2.2.0",
+        "rollup": "^0.25.8",
+        "rollup-plugin-babel": "^2.4.0",
+        "rollup-plugin-uglify": "^0.3.1"
+    },
+    "dependencies": {
+        "uaop": "0.0.1-alpha.2"
+    },
+    "edp": {
+        "main": "main"
+    }
 }
